Move receiver and service to their own folders
